requirements.txt
discord.py
pymongo
aiohttp
python-dotenv
requests